Rename App to Hometown. There is no default App now.
