+++ hometown/src/Hometown.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { render, screen } from '@testing-library/react'
-import './App'
+import './Hometown'
 
 declare global {
   namespace JSX {
