+++ hometown/src/index.tsx
@@ -1,1 +1,1 @@
-import './App'
+import './Hometown'
